refactor(backoffice-v2): moved useNavigateBack to a file of its own
