+++ apps/backoffice-v2/src/common/hooks/useNavigateBack/useNavigateBack.tsx
@@ -1,0 +1,11 @@
+import { useNavigate } from 'react-router-dom';
+import { useCallback } from 'react';
+
+export const useNavigateBack = () => {
+  const navigate = useNavigate();
+  const navigateBack = useCallback(() => {
+    navigate(-1);
+  }, [navigate]);
+
+  return navigateBack;
+};
