+++ apps/backoffice-v2/src/pages/TransactionMonitoringAlertsAnalysis/TransactionMonitoringAlertsAnalysis.page.tsx
@@ -1,31 +1,27 @@
 import { AlertAnalysisSheet } from '@/pages/TransactionMonitoringAlerts/components/AlertAnalysisSheet';
 import { useTransactionsQuery } from '@/domains/transactions/hooks/queries/useTransactionsQuery/useTransactionsQuery';
 import { useSerializedSearchParams } from '@/common/hooks/useSerializedSearchParams/useSerializedSearchParams';
+import { useNavigateBack } from '@/common/hooks/useNavigateBack/useNavigateBack';
 import { useNavigate } from 'react-router-dom';
-import { useCallback } from 'react';
-
-export const useNavigateBack = () => {
-  const navigate = useNavigate();
-  const navigateBack = useCallback(() => {
-    navigate(-1);
-  }, [navigate]);
-
-  return navigateBack;
-};
 
 export const TransactionMonitoringAlertsAnalysisPage = () => {
   const [{ businessId }] = useSerializedSearchParams();
   const { data: transactions } = useTransactionsQuery({
     businessId: businessId ?? '',
   });
-  const isSheetOpen = true;
   const navigateBack = useNavigateBack();
+  const navigate = useNavigate();
+  const onNavigateBack = () => {
+    if (window.history.state.idx <= 1) {
+      navigate('../');
+
+      return;
+    }
+
+    navigateBack();
+  };
 
   return (
-    <AlertAnalysisSheet
-      transactions={transactions ?? []}
-      isOpen={isSheetOpen}
-      onOpenStateChange={navigateBack}
-    />
+    <AlertAnalysisSheet transactions={transactions ?? []} onOpenStateChange={onNavigateBack} />
   );
 };
